reading pngs :D

+++ RaveEngine/RaveEngine/Libraries/zlib/gzwrite.c
@@ -208,8 +208,8 @@ local z_size_t gz_write(state, buf, len)
             have = (unsigned)((state->strm.next_in + state->strm.avail_in) -
                               state->in);
             copy = state->size - have;
-            if (copy > len)
-                copy = len;
+            if (copy > (unsigned int)len)
+                copy = (unsigned int)len;
             memcpy(state->in + have, buf, copy);
             state->strm.avail_in += copy;
             state->x.pos += copy;
@@ -228,8 +228,8 @@ local z_size_t gz_write(state, buf, len)
         state->strm.next_in = (z_const Bytef *)buf;
         do {
             unsigned n = (unsigned)-1;
-            if (n > len)
-                n = len;
+            if (n > (unsigned int)len)
+                n = (unsigned int)len;
             state->strm.avail_in = n;
             state->x.pos += n;
             if (gz_comp(state, Z_NO_FLUSH) == -1)
@@ -368,7 +368,7 @@ int ZEXPORT gzputs(file, str)
 
     /* write string */
     len = strlen(str);
-    ret = gz_write(state, str, len);
+    ret = (int)gz_write(state, str, len);
     return ret == 0 && len != 0 ? -1 : ret;
 }
